Replace os.sep with "/" for js script path
A user reported that leafelt maps generated by this plugin doesn't work on firefox. The cause of this was that if the index.html was generated on a windows machine, it has a separator '\' instead of '/'. Firefox fails to load this JS file since it can't interpret the path. While using os.sep is correct for system paths, for javascript paths, using forward slash would be better.